修正：考试成绩查询的得分区间判断，当查询不及极成绩时，之前是按0-59分作为查询区间，会漏掉59-60分之间的成绩

+++ Sourcecode/Song.ServiceImpls/ExaminationCom.cs
@@ -1288,7 +1288,7 @@ namespace Song.ServiceImpls
             return Gateway.Default.Count<ExamResults>(ExamResults._.Exam_ID == examid);
             
         }
-        public ExamResults[] Results(int examid, string name, string idcard, int min, int max, int size, int index, out int countSum)
+        public ExamResults[] Results(int examid, string name, string idcard, float min, float max, int size, int index, out int countSum)
         {
             WhereClip wc = ExamResults._.Exam_ID == examid;           
             if (min >= 0) wc.And(ExamResults._.Exr_ScoreFinal >= min);
